RDKBACCL-1381 : Cherry-pick the scarthgap changes to rdkb-bpi-2025q4 branch

+++ meta-rdk-mtk-bpir4/recipes-kernel/linux-firmware/linux-firmware_%.bbappend
@@ -1,7 +1,9 @@
 #Remove duplicate package installation in populate sdk
-do_install_append() {
+do_install:append() {
+        if ${@bb.utils.contains('DISTRO_FEATURES','krikstone','true','false',d)}; then
                 rm -rf ${D}${base_libdir}/firmware/mediatek/mt7996
                 rm -rf ${D}${base_libdir}/firmware/mediatek/mt7988
                 rm -rf ${D}${base_libdir}/firmware/mediatek/mt7996/mt7996*
+        fi
 }
